Create authorization service and minor fixes
